slider and date selection added to the search criteria ui
