+++ Khoji/ViewModels/SearchSettings.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct SearchSettings {
+    var searchByDate: Bool = false
+    var searchBySize: Bool = false
+}
